+++ src/echo/__init__.py
@@ -1,3 +1,3 @@
 """Echo programming language."""
 
-__version__ = "0.8.4"
+__version__ = "0.8.5"